Move Space type to types folder
