+++ ui/src/Space/SpaceClient.test.tsx
@@ -18,7 +18,7 @@
 import Axios from 'axios';
 import SpaceClient from './SpaceClient';
 import Cookies from 'universal-cookie';
-import {createEmptySpace} from './Space';
+import {createEmptySpace} from 'Types/Space';
 import {MatomoWindow} from '../CommonTypes/MatomoWindow';
 import TestData from '../Utils/TestData';
 import {UserSpaceMapping} from './UserSpaceMapping';
