+++ ui/src/Tags/TagClient.interface.ts
@@ -18,7 +18,7 @@
 import {AxiosResponse} from 'axios';
 import {TagInterface} from './Tag.interface';
 import {TagRequest} from './TagRequest.interface';
-import {Space} from '../Space/Space';
+import {Space} from 'Types/Space';
 
 export interface TagClient {
      get(spaceUuid: string): Promise<AxiosResponse<TagInterface[]>>;
